Insert add/sub rsp, x instructions when function does not have
And there is a memory access.
The issue is that when the function preambel is like this:
push    rbp
mov     rbp, rsp
mov     [rbp-0x14], edi
mov     [rbp-0x18], esi

Then the stack is used but the rsp is not updated.
The push reg instructions created for the instrumentation then overwrite
the values living on the stack.

Failing example: problems/p1 in the chk function